config/pyproject.toml
[project]
name = "config"
version = "0.1.0"
description = ""
authors = [
    {name = "rhusiev",email = "63148080+rhusiev@users.noreply.github.com"}
]
license = {text = "GPLv3"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)",
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
